refactor!: move config, cursor & netrw files
Move the 'config.lua', 'cursor.lua' and 'netrw.lua' into a dedicated
"config" directory to better structure the source code

This is breaking because of the 'cursor' and 'netrw' file now need new
require statements, but since these should only been used for internal
stuff this should not be a big problem
