+++ lua/drex/config/cursor.lua
